VASP 6.3.0 for LUMI/22.06
Added the libxc plug-in and verified that it is active. The VASP test suite passes, and the performance seems unchanged w.r.t to a simple scale-out test of a big supercell test on 1-12 compute nodes. It was not possible to test any larger jobs currently, due to the network problems on LUMI post-upgrade.