Update spglib-1.16.3-cpeGNU-22.08.eb
Usage on CP2K EB recipe (cpeGNU 22.08) spglib was broken as it couldn't file the tarball. Upon further investigation it turns out that the url links to where spglib was being deployed had changed. The EB recipe has been updated to reflect this. 

Further testing with new EB in progress