Fixes for VASP IO problems during startup
VASP 5 and 6 (with HDF5 enabled) can hang for several minutes on LUMI. The reason is that all ranks try to read the input files with write access at the same time. These EB recipes applies patches to the VASP sources which reads POTCAR and INCAR in read-only mode. It helps.